Add docker and yarn to the dev-workspace
Start fixing the test scripts.
