+++ .github/ISSUE_TEMPLATE/release-pro-plugin.md
@@ -15,11 +15,11 @@ To release the plugin please make sure to check all the checkboxes below.
 - [ ] Make sure to directly merge or use Pull Requests to merge hotfixes or features branches into the release branch.
 - [ ] Enter the `dev-workspace` folder and run the command `/run` to enter to the continer. The next commands should run there.
 - [ ] Run `$ composer update --dry-run` and check if there is any relevant update. Check if you need to lock the current version for any dependency. No dev requirement should be added to the git repository. If any update is needed, run `$ composer update` and commit the changes.
-- [ ] Build JS files to production running `$ npm run build-js` and commit.
+- [ ] Build JS files to production running `$ yarn run build-js` and commit.
 - [ ] Run WP VIP scan to make sure no warnings or errors > 5 exists: `$ vendor/bin/phpcs`.
 - [ ] Update the changelog - make sure all the changes are there with a user-friendly description and that the release date is correct.
 - [ ] Update the version number to the next stable version in the main plugin file and `readme.txt`. Commit the changes to the release branch.
-- [ ] Build the zip package, running `$ npm run build`. It should create a package in the `./dist` dir.
+- [ ] Build the zip package, running `$ yarn run build`. It should create a package in the `./dist` dir.
 - [ ] Send the new package to the team for testing.
 
 ### Release Checklist
