+++ .github/workflows/deploy-wp.org.yml
@@ -23,7 +23,7 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Build the package
-        run: npm run build
+        run: yarn run build
 
       - name: WordPress.org plugin deploy
         uses: 10up/action-wordpress-plugin-deploy@stable
@@ -34,7 +34,7 @@ jobs:
           BUILD_DIR: /project/dist/post-expirator
 
       - name: Save version number to version.txt
-        run: npm run get-version
+        run: yarn run get-version
 
       - name: Get version number from version.txt
         id: get-version
