+++ README-build.md
@@ -10,20 +10,20 @@ builder/docker/scripts/build build
 builder/docker/scripts/build build-dir
 ```
 
-If you prefer you can run the npm scripts:
+If you prefer you can run the js scripts:
 
 ```bash
-npm run build
-npm run build:dir
-npm run build:clean
+yarn run build
+yarn run build:dir
+yarn run build:clean
 ```
 
 ## Build Assets
 
 ```bash
-npm run jsbuild
-npm run jsbuild:dev
-npm run jsbuild:watch
+yarn run jsbuild
+yarn run jsbuild:dev
+yarn run jsbuild:watch
 ```
 
 ## Tests
@@ -36,18 +36,18 @@ tests/bin/tests php5.6 stop
 tests/bin/tests php5.6 run tests/codeception/acceptance/features/settings.feature
 ```
 
-If you prefer you can run the npm scripts:
+If you prefer you can run the yarn scripts:
 
 ```bash
-npm run tests:start
-npm run tests:stop
-npm run tests:run unit
-npm run tests:run wordpress
-npm run tests:run acceptance
-npm run tests:run tests/codeception/acceptance/features/settings.feature
-npm run tests:build:run wordpress
-npm run tests:build:run acceptance
-npm run tests:build:run tests/codeception/acceptance/features/settings.feature
+yarn run tests:start
+yarn run tests:stop
+yarn run tests:run unit
+yarn run tests:run wordpress
+yarn run tests:run acceptance
+yarn run tests:run tests/codeception/acceptance/features/settings.feature
+yarn run tests:build:run wordpress
+yarn run tests:build:run acceptance
+yarn run tests:build:run tests/codeception/acceptance/features/settings.feature
 ```
 
 The command `tests:build:run` run the PHP build process before running the tests. Use that if you are testing code that was recently modified. If you wan't to just repeat the tests without building (when no change was done in the plugin code) you can add a `-` to the end of the command: `tests:run`.
@@ -57,13 +57,13 @@ After `tests:build:run` or `tests:run` you can pass any codeception argument you
 ## Running WP CLI on devkinsta
 
 ```bash
-npm run wp
+yarn run wp
 ```
 
-You can pass arguments after the command, normally. But for passing named arguments you need to pass the `--` argument before them, otherwise those named arguments won't bypass and `npm` will interpret them instead. For example:  
+You can pass arguments after the command, normally. But for passing named arguments you need to pass the `--` argument before them, otherwise those named arguments won't bypass and `npm` will interpret them instead. For example:
 
 ```bash
-npm run wp cron event run -- --due-now
+yarn run wp cron event run -- --due-now
 ```
 
 ## TODO
