+++ package.json
@@ -2,7 +2,7 @@
   "name": "post-expirator",
   "version": "2.7.0",
   "description": "post-expirator",
-  "type": "commonjs",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/publishpress/PublishPress-Future.git"
@@ -53,6 +53,8 @@
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
     "braces": "^3.0.2",
+    "child_process": "^1.0.2",
+    "inquirer": "^9.1.5",
     "inquirer-autocomplete-prompt": "^3.0.0",
     "inquirer-interrupted-prompt": "^2.0.1",
     "node-ray": "^1.19.4",
