+++ tests/bin/prompt-version.js
@@ -1,29 +1,18 @@
-#!/usr/bin/env node
+#!/usr/bin/env node --experimental-modules
 
 /*
  * Copyright (c) 2022. PublishPress, All rights reserved.
  */
-import inquirer from "inquirer";
 
-import {execSync} from 'child_process';
-const exec = commands => {
-    execSync(commands, {stdio: 'inherit', shell: true});
-};
+import { promptUser, executeCommand } from './modules/prompt-user.mjs';
 
 const additionalArgs = process.argv.slice(2);
 
-inquirer
-    .prompt([
-        {
-            type: "list",
-            name: "version",
-            message: "Select a PHP version",
-            choices: ["5.6", "7.4", "8.0", "8.1", "8.2"]
-        },
-    ])
+promptUser()
     .then((answers) => {
-        const command = additionalArgs.join(' ').replace('{{VERSION}}', answers.version);
-        exec(command);
+        const command = additionalArgs.join(' ');
+        executeCommand(command, answers.version);
     })
     .catch((error) => {
+        console.error(error);
     });
